Extract NumPy training helpers for test isolation
Move fill_arange_position_ids and sdpa_causal_mask_bool_fortran_fill to
examples/numpy_helpers.py so test_causal_mask does not exec_module
gpt2_training (which imports the nntile extension).
